leaderboard_space_submission/requirements.txt
gradio>=5.0.0
pandas>=2.0.0
pyyaml>=6.0.0
